+++ rust/operator/src/discovery.rs
@@ -44,15 +44,24 @@ pub async fn build_discovery_configmaps(
     owner: &impl Resource<DynamicType = ()>,
     kafka: &KafkaCluster,
     svc: &Service,
+    app_managed_by: &str,
 ) -> Result<Vec<ConfigMap>, Error> {
-    let name = owner.name();
+    let name = owner.name_unchecked();
+    let port_name = kafka.client_port_name();
     Ok(vec![
-        build_discovery_configmap(&name, owner, kafka, service_hosts(svc, "kafka")?)?,
+        build_discovery_configmap(
+            &name,
+            owner,
+            kafka,
+            service_hosts(svc, port_name)?,
+            app_managed_by,
+        )?,
         build_discovery_configmap(
             &format!("{}-nodeport", name),
             owner,
             kafka,
-            nodeport_hosts(client, svc, "kafka").await?,
+            nodeport_hosts(client, svc, port_name).await?,
+            app_managed_by,
         )?,
     ])
 }
@@ -65,6 +74,7 @@ fn build_discovery_configmap(
     owner: &impl Resource<DynamicType = ()>,
     kafka: &KafkaCluster,
     hosts: impl IntoIterator<Item = (impl Into<String>, u16)>,
+    app_managed_by: &str,
 ) -> Result<ConfigMap, Error> {
     // Write a list of bootstrap servers in the format that Kafka clients:
     // "{host1}:{port1},{host2:port2},..."
@@ -88,6 +98,7 @@ fn build_discovery_configmap(
                     kafka
                         .image_version()
                         .context(KafkaVersionParseFailureSnafu)?,
+                    app_managed_by,
                     &KafkaRole::Broker.to_string(),
                     "discovery",
                 )