+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Added default resource requests (memory and cpu) for Kafka pods. ([#485])
+- Support for Kafka 3.3.1 ([#492])
+- Orphaned resources are deleted ([#495])
+
+[#485]: https://github.com/stackabletech/kafka-operator/pull/485
+[#492]: https://github.com/stackabletech/kafka-operator/pull/492
+[#495]: https://github.com/stackabletech/kafka-operator/pull/495
+
+### Changed
+
+- Change port names from `http`/`https` to `kafka`/`kafka-tls` ([#472]).
+
+[#472]: https://github.com/stackabletech/kafka-operator/pull/472
+
+## [0.7.0] - 2022-09-06
+
+### Added
+
+- BREAKING: TLS encryption and authentication support for internal and client communications. This is breaking for clients because the cluster is secured per default, which results in a client port change ([#442]).
+
 ### Changed
 
 - operator-rs: 0.21.1 -> 0.22.0 ([#430]).
@@ -14,6 +36,7 @@ All notable changes to this project will be documented in this file.
 [#430]: https://github.com/stackabletech/kafka-operator/pull/430
 [#431]: https://github.com/stackabletech/kafka-operator/pull/431
 [#434]: https://github.com/stackabletech/kafka-operator/pull/434
+[#442]: https://github.com/stackabletech/kafka-operator/pull/442
 
 ## [0.6.0] - 2022-06-30